+++ tor_async_util/jsonschemas.py
@@ -13,3 +13,5 @@ def _load_jsonschema(schema_name):
         return json.load(fp)
 
 get_noop_response = _load_jsonschema('get_noop_response')
+
+get_health_response = _load_jsonschema('get_health_response')