+++ tor_async_util/tests/unit_tests.py
@@ -1292,3 +1292,70 @@ class NoOpTestCase(RequestHandlerTestCase):
         response = self.fetch(TestNoOpRequestHandler.url_spec, method='GET')
         self.assertEqual(response.code, httplib.OK)
         self.assertNoDebugDetail(response)
+
+
+class AsyncHealthCheck(object):
+
+    def __init__(self, is_quick, async_state=None):
+        object.__init__(self)
+
+        self.is_quick = is_quick
+        self.async_state = async_state
+
+    def check(self, callback):
+        if self.is_quick:
+            details = None
+        else:
+            details = {
+                'other_service': True,
+            }
+
+        callback(True, details, self)
+
+
+class HealthCheckRequestHandler(tor_async_util.RequestHandler):
+
+    url_spec = r'/_health'
+
+    @tornado.web.asynchronous
+    def get(self):
+        tor_async_util.generate_health_check_response(self, AsyncHealthCheck)
+
+
+class HealthCheckTestCase(RequestHandlerTestCase):
+    """Unit tests for generate_health_check_response()."""
+
+    def get_app(self):
+        handlers = [
+            (
+                HealthCheckRequestHandler.url_spec,
+                HealthCheckRequestHandler
+            ),
+        ]
+        return tornado.web.Application(handlers=handlers)
+
+    def test_happy_path_no_is_quick(self):
+        response = self.fetch(HealthCheckRequestHandler.url_spec, method='GET')
+        self.assertEqual(response.code, httplib.OK)
+        self.assertNoDebugDetail(response)
+
+    def test_happy_path_is_quick_equals_true(self):
+        response = self.fetch('%s?quick=true' % HealthCheckRequestHandler.url_spec, method='GET')
+        self.assertEqual(response.code, httplib.OK)
+        self.assertNoDebugDetail(response)
+
+    def test_happy_path_is_quick_equals_false(self):
+        response = self.fetch('%s?quick=false' % HealthCheckRequestHandler.url_spec, method='GET')
+        self.assertEqual(response.code, httplib.OK)
+        self.assertNoDebugDetail(response)
+
+    def test_is_quick_equals_invalid_value(self):
+        response = self.fetch('%s?quick=davewashere' % HealthCheckRequestHandler.url_spec, method='GET')
+        self.assertEqual(response.code, httplib.BAD_REQUEST)
+        self.assertDebugDetail(response, tor_async_util.HEALTH_CHECK_GDD_INVALID_QUICK_ARGUMENT)
+
+    def test_bad_response_body(self):
+        with WriteAndVerifyPatcher(is_ok=False):
+            response = self.fetch(HealthCheckRequestHandler.url_spec, method='GET')
+            self.assertEqual(response.code, httplib.INTERNAL_SERVER_ERROR)
+            self.assertDebugDetail(response, tor_async_util.HEALTH_CHECK_GDD_INVALID_RESPONSE_BODY)